feat: Move TextComponent to toolbox
